Merging the file input support with the manifest directory walk method.

+++ payu/models/um.py
@@ -11,6 +11,7 @@ from __future__ import print_function
 # Standard Library
 import datetime
 import fileinput
+import glob
 import imp
 import os
 import shutil
@@ -18,6 +19,7 @@ import string
 
 # Extensions
 import f90nml
+import yaml
 
 # Local
 from payu.fsops import mkdir_p, make_symlink
@@ -32,6 +34,7 @@ class UnifiedModel(Model):
 
         self.model_type = 'um'
         self.default_exec = 'um'
+        self.restart_calendar_file = self.model_type + '.res.yaml'
 
         # TODO: many of these can probably be ignored.
         self.config_files = ['CNTLALL', 'prefix.CNTLATM', 'prefix.CNTLGEN',
@@ -40,7 +43,8 @@ class UnifiedModel(Model):
                              'hnlist', 'ihist', 'INITHIS',
                              'namelists', 'PPCNTL', 'prefix.PRESM_A',
                              'SIZES', 'STASHC', 'UAFILES_A', 'UAFLDS_A',
-                             'parexe', 'cable.nml']
+                             'parexe', 'cable.nml', 'um_env.py']
+        self.optional_config_files = ['input_atm.nml']
 
         self.restart = 'restart_dump.astart'
 
@@ -50,6 +54,16 @@ class UnifiedModel(Model):
         self.work_input_path = os.path.join(self.work_path, 'INPUT')
 
     def archive(self):
+        super(UnifiedModel, self).archive()
+
+        # Delete all the stdout log files except the root PE
+        # Sorts to ensure root PE is first entry
+        files = sorted(glob.glob(
+                       os.path.join(self.work_path, 'atm.fort6.pe*')),
+                       key=lambda name: int(name.rpartition('.')[-1][2:]))
+        if len(files) > 1:
+            for f_path in files[1:]:
+                os.remove(f_path)
 
         mkdir_p(self.restart_path)
 
@@ -59,23 +73,44 @@ class UnifiedModel(Model):
 
         resubmit_inc = nml['NLSTCALL']['RUN_RESUBMIT_INC']
         runtime = um_time_to_time(resubmit_inc)
-        runtime = datetime.timedelta(seconds=runtime)
+        # runtime = datetime.timedelta(seconds=runtime)
 
         basis_time = nml['NLSTCALL']['MODEL_BASIS_TIME']
         init_date = um_date_to_date(basis_time)
 
-        end_date = date_to_um_dump_date(init_date + runtime)
+        end_date = cal.date_plus_seconds(init_date,
+                                         runtime,
+                                         cal.GREGORIAN)
+
+        # Save model time to restart next run
+        with open(os.path.join(self.restart_path,
+                  self.restart_calendar_file), 'w') as restart_file:
+            restart_file.write(yaml.dump({'end_date': end_date},
+                               default_flow_style=False))
+
+        end_date = date_to_um_dump_date(end_date)
 
         restart_dump = os.path.join(self.work_path,
                                     'aiihca.da{0}'.format(end_date))
         f_dst = os.path.join(self.restart_path, self.restart)
-        shutil.copy(restart_dump, f_dst)
+        if os.path.exists(restart_dump):
+            shutil.copy(restart_dump, f_dst)
+        else:
+            print('payu: error: Model has not produced a restart dump file:\n'
+                  '{} does not exist.\n'
+                  'Check DUMPFREQim in namelists'.format(restart_dump))
 
     def collate(self):
         pass
 
     def setup(self):
         super(UnifiedModel, self).setup()
+
+        # Set up environment variables needed to run UM.
+        # Look for a python file in the config directory.
+        um_env = imp.load_source('um_env',
+                                 os.path.join(self.control_path, 'um_env.py'))
+        um_vars = um_env.vars
 
         # Stage the UM restart file.
         if self.prior_restart_path and not self.expt.repeat_run:
@@ -84,14 +119,9 @@ class UnifiedModel(Model):
 
             if os.path.isfile(f_src):
                 make_symlink(f_src, f_dst)
-
-        # Set up environment variables needed to run UM.
-        # Look for a python file in the config directory.
-        um_env = imp.load_source('um_env',
-                                 os.path.join(self.control_path, 'um_env.py'))
-        um_vars = um_env.vars
-
-        assert len(self.input_paths) == 1
+                # every run is an NRUN with an updated ASTART file
+                um_vars['ASTART'] = self.restart
+                um_vars['TYPE'] = 'NRUN'
 
         # Set paths in environment variables.
         for k in um_vars.keys():
@@ -110,20 +140,17 @@ class UnifiedModel(Model):
         work_nml_path = os.path.join(self.work_path, 'namelists')
         work_nml = f90nml.read(work_nml_path)
 
+        restart_calendar_path = os.path.join(self.work_init_path,
+                                             self.restart_calendar_file)
+
         # Modify namelists for a continuation run.
-        if self.prior_output_path and not self.expt.repeat_run:
+        if self.prior_restart_path and not self.expt.repeat_run \
+           and os.path.exists(restart_calendar_path):
 
-            prior_nml_path = os.path.join(self.prior_output_path, 'namelists')
-            prior_nml = f90nml.read(prior_nml_path)
+            with open(restart_calendar_path, 'r') as restart_file:
+                restart_info = yaml.load(restart_file)
 
-            basis_time = prior_nml['NLSTCALL']['MODEL_BASIS_TIME']
-            init_date = um_date_to_date(basis_time)
-            resubmit_inc = prior_nml['NLSTCALL']['RUN_RESUBMIT_INC']
-            runtime = um_time_to_time(resubmit_inc)
-
-            run_start_date = cal.date_plus_seconds(init_date,
-                                                   runtime,
-                                                   cal.GREGORIAN)
+            run_start_date = restart_info['end_date']
 
             # Write out and save new calendar information.
             run_start_date_um = date_to_um_date(run_start_date)
@@ -168,7 +195,7 @@ def date_to_um_dump_date(date):
 
     assert(date.month <= 12)
 
-    decade = date.year / 10
+    decade = date.year // 10
     # UM can only handle 36 decades then goes back to the beginning.
     decade = decade % 36
     year = date.year % 10
@@ -231,4 +258,4 @@ def time_to_um_time(seconds):
 
     assert(seconds % 86400 == 0)
 
-    return [0, 0, seconds / 86400, 0, 0, 0]
+    return [0, 0, seconds // 86400, 0, 0, 0]
